Added documentation to timer and removed timer_setup()
timer_setup() should not be exposed in the header file, since we
do not want other functions other than timer functions calling it.
I also added documentation to the timer functions so that other
people can use it! I really should put these in a util folder for
future projects.